`elisp-def--show-occurrence': support C functions
When jumping to functions defined in C, `elisp-def' errors, without providing
any of the features it should (blinking, putting the cursor on the symbol name
...). Fix this, supporting functions defined in C.

This is done by checking for `c-mode' and `c++-mode' as `major-mode' and
considering the text between the first double-quoted string as the symbol name.
Blinking is supported.